pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["mauricio8989 <mauriciosouza8989@gmail.com>"]
readme = "README.md"
package-mode=false

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1"
django-rest-framework = "^0.1.0"
factory-boy = "^3.3.1"
pytest = "^8.3.2"
pytest-django = "^4.9.0"
djangorestframework = "^3.15.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
factory-boy = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
